Give up on reflection, just match on class name
Unfortunately the reflection approach didn't work if we publish using
1.3.13 and run against 1.4.0.

It throws a ClassCastException as soon as we do
`analysis.relations.allLibraryDeps.map(...)`, because the type signature
of `allLibraryDeps` it was built against (`Set[File]`) doesn't match the
signature of `allLibraryDeps` on the classpath (`Set[VirtualFileRef]`).

Worked around this by casting to `Set[AnyRef]`, but that means we can't
use a type tag anymore. So I gave up on reflection and used `getClass`
instead. Pretty nasty but it works.

Confirmed manually that if we publish using 1.4.0 we can only run
against 1.4.0, so we need to publish using 1.3.13.