Handle semver metadata in version string
Update how version is parsed for pom files in the jar file
to dependency logic to to cater for characters that are valid
in version strings but encoded in the file path.
For example, "+" which is used for semver metadata, encoded as
"%2B".

Before this change 7.2.27+33-04a1ea9e-SNAPSHOT version string
was not handled consistently and would result in different
values in the declared/used compile dependencies.

Fixes: https://github.com/cb372/sbt-explicit-dependencies/issues/110